+++ src/Input/Input.spec.js
@@ -4,6 +4,7 @@ import sinon from 'sinon';
 import inputDriverFactory from './Input.driver';
 import Input from '.';
 import { createDriverFactory } from 'wix-ui-test-utils/driver-factory';
+import { render, cleanup } from '../../test/utils/react';
 import { inputTestkitFactory, tooltipTestkitFactory } from '../../testkit';
 import { inputTestkitFactory as enzymeInputTestkitFactory } from '../../testkit/enzyme';
 import {
@@ -106,23 +107,26 @@ describe('Input', () => {
       const props = {
         type: 'text',
         name: 'gal',
-        onChange
+        onChange,
       };
-      const driver = createDriver(<Input {...props}/>);
+      const driver = createDriver(<Input {...props} />);
       driver.enterText('some text');
       const eventTarget = onChange.mock.calls[0][0].target;
-      expect(eventTarget).toEqual({name: 'gal', type: 'text', value: 'some text'});
+      expect(eventTarget).toEqual({
+        name: 'gal',
+        type: 'text',
+        value: 'some text',
+      });
     });
   });
-
 
   describe('name attribute', () => {
     it('should pass down to the wrapped input', () => {
       const props = {
-        name: 'hello'
+        name: 'hello',
       };
 
-      const driver = createDriver(<Input {...props}/>);
+      const driver = createDriver(<Input {...props} />);
       expect(driver.getName()).toEqual(props.name);
     });
   });
@@ -130,10 +134,10 @@ describe('Input', () => {
   describe('type attribute', () => {
     it('should pass down to the wrapped input', () => {
       const props = {
-        type: 'number'
+        type: 'number',
       };
 
-      const driver = createDriver(<Input {...props}/>);
+      const driver = createDriver(<Input {...props} />);
       expect(driver.getType()).toEqual(props.type);
     });
   });
@@ -714,8 +718,41 @@ describe('Input', () => {
   });
 });
 
+describe('testkit exists', () => {
+  afterEach(() => cleanup());
+
+  it('should NOT exist', () => {
+    const dataHook1 = 'hook1';
+    const dataHook2 = 'hook2';
+    const value = 'hello';
+    const onChange = () => {};
+
+    const { container } = render(
+      <Input value={value} onChange={onChange} dataHook={dataHook1} />,
+    );
+
+    const driver = inputTestkitFactory({
+      wrapper: container,
+      dataHook: dataHook2,
+    });
+
+    expect(driver.exists()).toBeFalsy();
+  });
+});
+
 describe('testkit', () => {
   it('should exist', () => {
+    const value = 'hello';
+    const onChange = () => {};
+    expect(
+      isTestkitExists(
+        <Input value={value} onChange={onChange} />,
+        inputTestkitFactory,
+      ),
+    ).toBe(true);
+  });
+
+  it('should NOT exist', () => {
     const value = 'hello';
     const onChange = () => {};
     expect(