+++ testkit/testkit-smoke.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { mount } from 'enzyme';
+import { render, cleanup } from 'react-testing-library';
 import path from 'path';
 
 import {
@@ -12,6 +13,8 @@ import importAllComponents from '../test/utils/import-all-components';
 
 import * as reactTestUtilsTestkitFactories from './index';
 import * as enzymeTestkitFactories from './enzyme';
+
+import Popover from '../src/Popover';
 
 /**
  * The following list ignores specified components and skip testkit tests for them.
@@ -59,9 +62,9 @@ const FAILING_COMPONENTS = [
   'Tabs',
   'TextArea',
   'TextField',
-  'Tooltip',
   'VBox', // Component has no testkit
   'Collapse',
+  'ContactItemBuilder',
 ];
 
 /**
@@ -90,6 +93,12 @@ const COMPONENTS = {
   },
   RichTextArea: {
     beforeAllHook: () => (window.getSelection = () => ({})),
+  },
+  Button: {
+    unidriver: true,
+    props: {
+      upgrade: true,
+    },
   },
   Tag: {
     props: {
@@ -152,10 +161,32 @@ const COMPONENTS = {
       columns: [{ title: 'A', render: row => row.a }],
     },
   },
+  Tooltip: {
+    props: {
+      content: "I'm the content",
+    },
+  },
   Modal: {
     props: {
       isOpen: false,
       contentLabel: 'modal_12345678',
+    },
+  },
+  Popover: {
+    props: {
+      children: [
+        <Popover.Element>
+          <div>I am the trigger!</div>
+        </Popover.Element>,
+        <Popover.Content>
+          <div>I am the content!</div>
+        </Popover.Content>,
+      ],
+    },
+  },
+  Proportion: {
+    props: {
+      children: <div />,
     },
   },
 };
@@ -218,6 +249,27 @@ const DRIVER_ASSERTS = {
             reactTestUtilsTestkitFactories[`${lowerFirst(name)}TestkitFactory`],
           ),
         );
+      });
+    });
+    describe('ReactTestUtils update dataHook', () => {
+      handleBeforeAllHook(beforeAllHook, afterAllHook);
+      /* eslint-disable jest/no-disabled-tests */
+      xit(`${name} should have an updated dataHook`, () => {
+        /* eslint-enable jest/no-disabled-tests */
+        const hook1 = 'my-data-hook-1';
+        const hook2 = 'my-data-hook-2';
+        const { rerender, container } = render(
+          React.createElement(component, { ...props, dataHook: hook1 }),
+        );
+        expect(
+          !!container.querySelector(`[data-hook="${hook1}"]`),
+        ).toBeTruthy();
+
+        rerender(React.createElement(component, { ...props, dataHook: hook2 }));
+        expect(
+          !!container.querySelector(`[data-hook="${hook2}"]`),
+        ).toBeTruthy();
+        cleanup();
       });
     });
   },